Refactor: 코드 1차 리팩토링
