+++ src/components/accom-search/calendar/AccomCalendar.component.jsx
@@ -1,12 +1,12 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { DateRange } from 'react-date-range';
 import { addDays } from 'date-fns';
+import { useAccomSearchStore } from '../../../states';
+import { formatDate } from '../../../utils/formatDate/formatDate';
+import { FaCalendar } from '../../../assets/icons/index';
 import './AccomCalendar.style.scss';
 import 'react-date-range/dist/styles.css';
 import 'react-date-range/dist/theme/default.css';
-import { FaCalendar } from '../../../assets/icons/index';
-import { useAccomSearchStore } from '../../../states';
-import { formatDate } from '../../../utils/formatDate/formatDate';
 
 const AccomCalendar = ({ ...props }) => {
   const { checkIn, checkOut, tripDay } = useAccomSearchStore((state) => state);
