+++ src/components/buttons/admin-icon-button/AdminIconButton.component.jsx
@@ -1,6 +1,4 @@
-import React from 'react';
 import './AdminIconButton.style.scss';
-import { useParams } from 'react-router-dom';
 
 const AdminIconButton = ({ onClick, children }) => {
   return (
