+++ src/pages/admin/accom-reg-form/local-components/room-list/RoomList.jsx
@@ -1,17 +1,11 @@
-import React from 'react';
-import './RoomList.style.scss';
-import { accomData } from '../../../../../assets/sample-data/accomSampleData';
-import {
-  BsFillHouseAddFill,
-  FaEdit,
-  MdDelete,
-} from '../../../../../assets/icons/index';
-import AdminIconButton from '../../../../../components/buttons/admin-icon-button/AdminIconButton.component';
 import { useNavigate, useParams } from 'react-router-dom';
+import { BsFillHouseAddFill } from '../../../../../assets/icons/index';
+import AdminIconButton from '../../../../../components/buttons/admin-icon-button/AdminIconButton.component';
 import AdminHeader from '../../../local-components/header/AdminHeader.component';
 import AdminSearch from '../../../local-components/header/search/AdminSearch.component';
 import AdminManagementList from '../../../local-components/list/AdminManagementList.component';
 import { roomData } from '../../../../../assets/sample-data/roomSampleData';
+import './RoomList.style.scss';
 
 const roomList = roomData;
 
@@ -63,34 +57,3 @@ const RoomList = () => {
 };
 
 export default RoomList;
-
-/*
-<div className='room-list'>
-          <table className='admin-table'>
-            <thead className='admin-table-header'>
-              <tr className='t-head-item'>
-                <th className='t-head-item-col col-w-10'>객실명</th>
-                <th className='t-head-item-col col-w-30'>객실수</th>
-                <th className='t-head-item-col col-w-30'>객실가격</th>
-                <th className='t-head-item-col col-w-20'>편집</th>
-                <th className='t-head-item-col col-w-10'>제거</th>
-              </tr>
-            </thead>
-            <tbody className='admin-table-body'>
-              {roomList.map((value, idx) => (
-                <tr key={idx} className='t-body-item' onClick={() => registerRoomPageHandler(value.room_sq)}>
-                  <td className='t-body-item-col'>{value.room_name}</td>
-                  <td className='t-body-item-col'>{value.room_count}</td>
-                  <td className='t-body-item-col'>{value.room_price.toLocaleString('ko-kr')}원</td>
-                  <td className='t-body-item-col edit-button'>
-                    <FaEdit />
-                  </td>
-                  <td className='t-body-item-col delete-button'>
-                    <MdDelete />
-                  </td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
-*/
