+++ src/pages/admin/reservation/ReservationManagementList.page.jsx
@@ -1,12 +1,10 @@
+import { useNavigate } from 'react-router-dom';
 import AdminHeader from '../local-components/header/AdminHeader.component';
 import AdminSearch from '../local-components/header/search/AdminSearch.component';
-import './ReservationManagementList.style.scss';
-import { accomData } from '../../../assets/sample-data/accomSampleData';
 import AdminManagementList from '../local-components/list/AdminManagementList.component';
 import { sampleReservations } from '../../../assets/sample-data/reservation';
-import { useNavigate } from 'react-router-dom';
-import { useState } from 'react';
-import { AdminPagination } from '../../../components/admin-pagination/AdminPagination.component';
+import './ReservationManagementList.style.scss';
+
 const dataList = sampleReservations;
 
 const reservationColumnList = [
@@ -21,8 +19,6 @@ const reservationColumnList = [
 const ReservationManagementList = () => {
   const navigate = useNavigate();
   const detailPageHandler = (resCode) => {
-    console.log('예약상세페이지로 이동!');
-    console.log(resCode);
     navigate(`/admin/reservations/${resCode}/detail`);
   };
