+++ src/pages/cart/CartMain.page.jsx
@@ -1,8 +1,8 @@
-import React from 'react';
-import './CartMain.style.scss';
 import CartBody from './body/CartBody.component';
 import CartHeader from './header/CartHeader.component';
 import { PageContainer } from '../../components';
+import './CartMain.style.scss';
+
 const CartMain = () => {
   return (
     <PageContainer className='cart-container'>
