+++ src/pages/cart/body/CartBody.component.jsx
@@ -1,11 +1,11 @@
-import React, { useEffect, useState } from 'react';
-import { GoCheckCircle, GoCheckCircleFill } from '../../../assets/icons/index';
-import PayInfo from '../../../components/pay/PayInfo.component';
-import './CartBody.style.scss';
-import { usePaymentInfoStore } from '../../../states';
-import Room from '../../../components/room-list/room/Room.component';
-import { accomData } from '../../../assets/sample-data/accomSampleData';
+import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import PayInfo from '../../../components/pay/PayInfo.component';
+import Room from '../../../components/room-list/room/Room.component';
+import { usePaymentInfoStore } from '../../../states';
+import { accomData } from '../../../assets/sample-data/accomSampleData';
+import { GoCheckCircle } from '../../../assets/icons/index';
+import './CartBody.style.scss';
 
 const accomDataList = accomData.accommodation_tb;
 
@@ -13,9 +13,7 @@ const CartBody = ({ className }) => {
   const navigate = useNavigate();
   const roomInfo = usePaymentInfoStore((state) => state.roomInfo);
   const { setRoomInfo } = usePaymentInfoStore((state) => state.actions);
-  const [selectedAll, setSelectedAll] = useState(false);
-
-  const [selectedList, setSelectedList] = useState([]);
+  const [setSelectedAll] = useState(false);
 
   // 전체 선택
   const checkAllHandler = () => {
@@ -47,33 +45,46 @@ const CartBody = ({ className }) => {
     navigate('/reservations', roomInfo);
   };
 
-  useEffect(() => {
-    // console.log(accomDataList);
-    // console.log(selectedList);
-    console.log(roomInfo);
-
-    // 상태 저장 후 예약 페이지 이동
-  }, [roomInfo]);
-
   return (
     <>
       <div className={className}>
         <div className='room-area__container'>
           <div className='all-check-area'>
-            <span className='check-all__true' onClick={checkAllHandler}>
+            <span
+              className='check-all__true'
+              onClick={checkAllHandler}
+            >
               <GoCheckCircle /> 전체선택
             </span>
-            <span className='check-all__false' onClick={checkAllClearHandler}>
+            <span
+              className='check-all__false'
+              onClick={checkAllClearHandler}
+            >
               전체해제
             </span>
           </div>
           {accomDataList?.map((value, idx) => {
             // 배열 안의 요소를 판별하여 boolean 값 전달
-            const isChecked = roomInfo.some((item) => item.accom_sq === value.accom_sq);
-            return <Room key={idx} className='room-item' value={value} checkArea={true} checkHandler={checkHandler} isChecked={isChecked} />;
+            const isChecked = roomInfo.some(
+              (item) => item.accom_sq === value.accom_sq
+            );
+            return (
+              <Room
+                key={idx}
+                className='room-item'
+                value={value}
+                checkArea={true}
+                checkHandler={checkHandler}
+                isChecked={isChecked}
+              />
+            );
           })}
         </div>
-        <PayInfo className='cart-pay-area__container' roomInfo={roomInfo} clickHandler={reserveHandler} />
+        <PayInfo
+          className='cart-pay-area__container'
+          roomInfo={roomInfo}
+          clickHandler={reserveHandler}
+        />
       </div>
     </>
   );
