+++ src/pages/main/local-components/bottom/MainBottom.component.jsx
@@ -1,6 +1,5 @@
-import React from 'react';
-import './MainBottom.style.scss';
 import LocationList from './location/LocationList.component';
+import './MainBottom.style.scss';
 
 const MainBottom = () => {
   return (
