+++ src/pages/main/local-components/bottom/card/Card.component.jsx
@@ -1,6 +1,5 @@
 import { Link } from 'react-router-dom';
 import { useAccomSearchStore } from '../../../../../states/accom-search/accomSearchStore';
-import { useEffect } from 'react';
 
 const Card = ({ cardText, ...props }) => {
   // 해당 카드 텍스트 값을 검색상태의 keyword 값으로 저장하기
