+++ src/pages/main/local-components/top/banner/Banner.component.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import './Banner.style.scss';
 
 const Banner = () => {
