+++ src/pages/non-member-reservation/box/NonMemberReservationBox.component.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
   ButtonPrimary,
   ButtonSecondary,
