+++ src/pages/payment/local-components/payment-info/body/PaymentInfoBody.component.jsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const PaymentInfoBody = ({ className }) => {
   return (
     <div className={className}>
