+++ src/states/payment-info/paymentInfoStore.js
@@ -1,5 +1,4 @@
 import { create } from 'zustand';
-import { persist } from 'zustand/middleware';
 
 const initialState = {
   resCode: '',
